requirements.txt
# dependencies
universal-intelligence[community]
beautifulsoup4>=4.12.0
pyyaml>=6.0.1
soupsieve>=2.5
requests>=2.31.0
